+++ avr/cores/megacommand/MCL/MCLClipBoard.h
@@ -10,10 +10,10 @@
 
 class MCLClipBoard {
 public:
-  uint8_t t_col;
-  uint8_t t_row;
-  uint8_t t_w;
-  uint8_t t_h;
+  int t_col;
+  int t_row;
+  int t_w;
+  int t_h;
 
   uint8_t copy_track;
 
@@ -30,8 +30,8 @@ public:
   bool paste_sequencer(uint8_t offset = 0);
   bool paste_sequencer_track(uint8_t source_track, uint8_t track);
 
-  bool copy(uint8_t col, uint8_t row, uint8_t w, uint8_t h);
-  bool paste(uint8_t col, uint8_t row);
+  bool copy(uint16_t col, uint16_t row, uint16_t w, uint16_t h);
+  bool paste(uint16_t col, uint16_t row);
 
 };
