+++ avr/cores/megacommand/MCL/SDDrivePage.cpp
@@ -107,7 +107,7 @@ void SDDrivePage::save_snapshot() {
   }
   //  Globals
   progress_max = 8;
-  for (uint8_t i = 0; i < 8; ++i) {
+  for (int i = 0; i < 8; ++i) {
     progress_i = i;
     mcl_gui.draw_progress("Saving global", i, 8);
     if (!MD.getBlockingGlobal(i)) {
@@ -121,7 +121,7 @@ void SDDrivePage::save_snapshot() {
   }
   //  Patterns
   progress_max = 128;
-  for (uint8_t i = 0; i < 128; ++i) {
+  for (int i = 0; i < 128; ++i) {
     progress_i = i;
     mcl_gui.draw_progress("Saving pattern", i, 128);
     DEBUG_PRINT(i);
@@ -136,7 +136,7 @@ void SDDrivePage::save_snapshot() {
   }
   //  Kits
   progress_max = 64;
-  for (uint8_t i = 0; i < 64; ++i) {
+  for (int i = 0; i < 64; ++i) {
     progress_i = 64;
     mcl_gui.draw_progress("Saving kit", i, 64);
     if (!MD.getBlockingKit(i)) {
@@ -149,7 +149,7 @@ void SDDrivePage::save_snapshot() {
     }
   }
   //  Songs
-  // for(uint8_t i=0;i<64;++i){
+  // for(int i=0;i<64;++i){
   // MD.getBlockingSong(i);
   // mcl_sd.write_data(&MD.song, sizeof(MD.song), &file); // <--- ??
   //}
@@ -207,7 +207,7 @@ void SDDrivePage::load_snapshot() {
 
   //  Globals
   progress_max = 8;
-  for (uint8_t i = 0; i < 8; ++i) {
+  for (int i = 0; i < 8; ++i) {
     progress_i = i;
     mcl_gui.draw_progress("Loading global", i, 8);
     if (!mcl_sd.read_data(&MD.global, sizeof(MD.global), &file)) {
@@ -225,7 +225,7 @@ void SDDrivePage::load_snapshot() {
   }
   //  Patterns
   progress_max = 128;
-  for (uint8_t i = 0; i < 128; ++i) {
+  for (int i = 0; i < 128; ++i) {
     progress_i = i;
     mcl_gui.draw_progress("Loading pattern", i, 128);
     if (!mcl_sd.read_data(&MD.pattern, sizeof(MD.pattern), &file)) {
@@ -240,7 +240,7 @@ void SDDrivePage::load_snapshot() {
   }
   //  Kits
   progress_max = 64;
-  for (uint8_t i = 0; i < 64; ++i) {
+  for (int i = 0; i < 64; ++i) {
     progress_i = i;
     mcl_gui.draw_progress("Loading kit", i, 64);
     if (!mcl_sd.read_data(&MD.kit, sizeof(MD.kit), &file)) {
